feat: adicionar listagem de bens
