+++ package.json
@@ -16,31 +16,32 @@
   "dependencies": {
     "@fortawesome/fontawesome-free": "^6.5.2",
     "@govbr-ds/core": "^3.5.1",
-    "@hookform/resolvers": "^3.6.0",
+    "@hookform/resolvers": "^3.9.0",
     "@million/lint": "1.0.0-rc.24",
-    "@tanstack/react-query": "^5.45.0",
-    "@tanstack/react-table": "^8.17.3",
+    "@tanstack/react-query": "^5.50.1",
+    "@tanstack/react-table": "^8.19.2",
     "@unocss/reset": "^0.60.4",
     "clsx": "^2.1.1",
     "comlink": "^4.4.1",
     "date-fns": "^3.6.0",
     "html2canvas": "^1.4.1",
-    "html2pdf.js": "^0.10.1",
+    "html2pdf.js": "^0.10.2",
+    "inbcm-xlsx-validator": "github:vadolasi/inbcm-xlsx-validator",
     "jspdf": "^2.5.1",
     "million": "^3.1.11",
     "msgpackr": "^1.10.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-dsgov": "^0.9.12",
-    "react-hook-form": "^7.51.5",
+    "react-hook-form": "^7.52.1",
     "react-hot-toast": "^2.4.1",
-    "react-router": "^6.23.1",
-    "react-router-dom": "^6.23.1",
-    "react-tooltip": "^5.27.0",
-    "tailwind-merge": "^2.3.0",
+    "react-router": "^6.24.1",
+    "react-router-dom": "^6.24.1",
+    "react-tooltip": "^5.27.1",
+    "tailwind-merge": "^2.4.0",
     "xlsx": "^0.18.5",
     "zod": "^3.23.8",
-    "zustand": "^4.5.2"
+    "zustand": "^4.5.4"
   },
   "devDependencies": {
     "@commitlint/cli": "^19.3.0",
@@ -51,8 +52,8 @@
     "@types/jspdf": "^2.0.0",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^7.13.0",
-    "@typescript-eslint/parser": "^7.13.0",
+    "@typescript-eslint/eslint-plugin": "^7.16.0",
+    "@typescript-eslint/parser": "^7.16.0",
     "@vitejs/plugin-basic-ssl": "^1.1.0",
     "@vitejs/plugin-react-swc": "^3.7.0",
     "@vitest/coverage-v8": "^1.6.0",
@@ -62,15 +63,15 @@
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-react-hooks": "^4.6.2",
-    "eslint-plugin-react-refresh": "^0.4.7",
+    "eslint-plugin-react-refresh": "^0.4.8",
     "husky": "^9.0.11",
     "jsdom": "^24.1.0",
     "lint-staged": "^15.2.7",
     "prettier": "^3.3.2",
-    "typescript": "^5.4.5",
+    "typescript": "^5.5.3",
     "unocss": "^0.60.4",
-    "vite": "^5.2.13",
-    "vite-plugin-pages": "^0.32.2",
+    "vite": "^5.3.3",
+    "vite-plugin-pages": "^0.32.3",
     "vite-plugin-wasm": "^3.3.0",
     "vite-plugin-wasm-pack": "^0.1.12",
     "vitest": "^1.6.0",
